python-training/requirements.txt
# PyTorch and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b0

# Market Data APIs
binance>=1.0.16
yfinance>=0.2.28

# Technical Analysis
ta-lib>=0.4.28  # Note: May require manual installation on some systems

# API and Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Database
supabase>=2.0.0
postgrest>=0.13.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
psutil>=5.9.0

# Logging and Monitoring
tensorboard>=2.15.0
wandb>=0.16.0  # Optional: For advanced experiment tracking

# Data Validation
marshmallow>=3.20.0

# Caching
redis>=5.0.0  # Optional: For distributed caching

# Testing (optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality (optional)
black>=23.0.0
flake8>=6.1.0
mypy>=1.6.0

# Notes:
# 1. For ta-lib installation issues, see: https://github.com/mrjbq7/ta-lib#installation
# 2. CUDA-enabled PyTorch: pip install torch --index-url https://download.pytorch.org/whl/cu118
# 3. For Apple Silicon: Use torch with MPS support
# 4. wandb and tensorboard are optional but highly recommended for experiment tracking
